railway_deployment/requirements.txt
fastapi
uvicorn
joblib
scikit-learn
ultralytics
torch
opencv-python-headless
requests
python-multipart
python-dotenv